+++ src/langserver/symbolsinformation/toolsGetNextObjectIdRequest.ts
@@ -1,0 +1,9 @@
+export class ToolsGetNextObjectIdRequest {
+    path: string | undefined;
+    objectType: string;
+    
+    constructor(newPath: string | undefined, newObjectType: string) {
+        this.path = newPath;
+        this.objectType = newObjectType;
+    }
+}