+++ src/langserver/symbolsinformation/toolsGetNextObjectIdResponse.ts
@@ -1,0 +1,3 @@
+export class ToolsGetNextObjectIdResponse {
+    id: number | undefined;
+}